done. Note: a client can only have one membership with a gym. Still needs work

+++ app/controllers/gyms_controller.rb
@@ -23,8 +23,12 @@ class GymsController < ApplicationController
     end
 
     def destroy
-        gym = Gym.find(params[:id])
-        gym.destroy
+        gym = Gym.find_by(id: params[:id])
+        if gym
+            gym.destroy
+        else
+            render json: { "error": "Gym not found" }, status: :not_found
+        end
     end
 
     private
